Assert on system scope permission based on supported flag
system scope is dropped from API policy in zed release
- https://review.opendev.org/c/openstack/glance/+/855435

and so does we addded some assert in glance-tempest-plugin rbac
tests for system scope permissions
- https://review.opendev.org/c/openstack/glance-tempest-plugin/+/855440

As glance-tempest-plugins should work on stable/xena onwards, we need to
make sure those system scope not allowed testing is done only
for stable/zed onwards where it is dropped.

This adds the config option and based on that assert on system scope
permissions. This config flag will be set as True for <=stable/yoga
so that these asserts can be skipped.

Change-Id: Ic6c09ced0eca03d0966154d3d7c6dfe1c38c3b13