glance-tempest-plugin branchless: add stable branch testing
Tempest and its plugins are branchless so that we can test master
and supported stable branches with same set of tests. This helps
to know if any backward incompatible changes merging in service side.

This makes the rbac jobs to use the master version of glance-tempest-plugin
for all supported table branch (stable/xena onwards are supported stable).

Also add jobs for stable branches so that we can capture any test changed
in backward incompatible way.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/882365
Depends-On: https://review.opendev.org/c/openstack/devstack/+/882366
Change-Id: Ic312ccd9700fadc1adb6ff6cc90e4397cb0b83f6